requirements.txt
mcp[cli]
fastapi
uvicorn
pydantic
pytest
python-dotenv
psutil
jinja2
ccxt
numpy
requests 